+++ src/core/PWSprefs.h
@@ -130,6 +130,7 @@ public:
     EnableWindowTransparency,
     ShowMenuSeparator,
     OptimizedCellSize,
+    ToolbarShowText,
     NumBoolPrefs};
 
   enum IntPrefs {Column1Width, Column2Width, Column3Width, Column4Width,