+++ src/ui/wxWidgets/MenuManageHandlers.cpp
@@ -63,6 +63,7 @@ void PasswordSafeFrame::OnPreferencesClick(wxCommandEvent& WXUNUSED(evt))
   PWSprefs* prefs = PWSprefs::GetInstance();
   bool showMenuSeprator = prefs->GetPref(PWSprefs::ShowMenuSeparator);
   bool optimizedCellSize = prefs->GetPref(PWSprefs::OptimizedCellSize);
+  bool toolbarShowText = prefs->GetPref(PWSprefs::ToolbarShowText);
   const StringX sxOldDBPrefsString(prefs->Store());
   OptionsPropertySheetDlg *window = new OptionsPropertySheetDlg(this, m_core);
   if (window->ShowModal() == wxID_OK) {
@@ -70,6 +71,14 @@ void PasswordSafeFrame::OnPreferencesClick(wxCommandEvent& WXUNUSED(evt))
       ReCreateMainToolbarSeparator(prefs->GetPref(PWSprefs::ShowMenuSeparator));
     if((optimizedCellSize != prefs->GetPref(PWSprefs::OptimizedCellSize)) && IsGridView() && IsShown())
       Show(true);
+    if(toolbarShowText != prefs->GetPref(PWSprefs::ToolbarShowText)) {
+      wxToolBar* tb = GetToolBar();
+      if(prefs->GetPref(PWSprefs::ToolbarShowText))
+        tb->SetWindowStyle(tb->GetWindowStyle() | wxTB_TEXT);
+      else
+        tb->SetWindowStyle(tb->GetWindowStyle() & ~wxTB_TEXT);
+      tb->Realize();
+    }
     
     StringX sxNewDBPrefsString(prefs->Store(true));
     // Update system tray icon if visible so changes show up immediately