+++ src/ui/wxWidgets/TreeCtrl.cpp
@@ -114,6 +114,7 @@ BEGIN_EVENT_TABLE( TreeCtrl, wxTreeCtrl )
   EVT_TREE_KEY_DOWN( ID_TREECTRL, TreeCtrl::OnKeyDown )
   EVT_TREE_BEGIN_DRAG(ID_TREECTRL, TreeCtrl::OnBeginDrag )
   EVT_TREE_END_DRAG(ID_TREECTRL, TreeCtrl::OnEndDrag )
+  EVT_MOTION( TreeCtrl::OnMouseMove )
   /*
     In Linux environments context menus appear on Right-Down mouse click.
     Which mouse click type (Right-Down/Right-Up) is the right one on a platform
@@ -136,6 +137,7 @@ BEGIN_EVENT_TABLE( TreeCtrl, wxTreeCtrl )
 END_EVENT_TABLE()
 
 const wchar_t GROUP_SEP = L'.';
+#define GROUP_SEL_STR wxT(".")
 
 // helper class to match CItemData with wxTreeItemId
 class PWTreeItemData : public wxTreeItemData
@@ -212,6 +214,7 @@ bool TreeCtrl::Create(wxWindow* parent, wxWindowID id, const wxPoint& pos, const
 TreeCtrl::~TreeCtrl()
 {
 ////@begin TreeCtrl destruction
+if(m_drag_image) wxDELETE(m_drag_image);
 ////@end TreeCtrl destruction
 }
 
@@ -224,8 +227,13 @@ void TreeCtrl::Init()
 ////@begin TreeCtrl member initialisation
   m_sort = TreeSortType::GROUP;
   m_show_group = false;
-  m_drag_item = NULL;
+  m_drag_item = nullptr;
   m_style = 0L;
+  m_lower_scroll_limit = m_upper_scroll_limit = 0;
+  m_scroll_timer.setOwner(this);
+  m_collapse_timer.setOwner(this);
+  m_drag_image = nullptr;
+  m_had_been_out = false;
 ////@end TreeCtrl member initialisation
 }
 
@@ -355,8 +363,6 @@ void TreeCtrl::UpdateGUI(UpdateGUICommand::GUI_Action ga, const pws_os::CUUID &e
  */
 void TreeCtrl::GUIRefreshEntry(const CItemData &item, bool WXUNUSED(bAllowFail))
 {
-  pws_os::Trace(wxT("TreeCtrl::GUIRefreshEntry"));
-
   if (item.GetStatus() == CItemData::ES_DELETED) {
     uuid_array_t uuid;
     item.GetUUID(uuid);
@@ -541,7 +547,7 @@ wxString TreeCtrl::GetPath(const wxTreeItemId &node) const
   for(iter = v.rbegin(); iter != v.rend(); iter++) {
     retval += *iter;
     if ((iter + 1) != v.rend())
-      retval += wxT(".");
+      retval += GROUP_SEL_STR;
   }
   return retval;
 }
@@ -556,7 +562,7 @@ wxString TreeCtrl::GetItemGroup(const wxTreeItemId& item) const
     if (path.IsEmpty())//parent is root
       return name; //group under root
     else
-      return path + wxT(".") + name; //sub-group of some (non-root) group
+      return path + GROUP_SEL_STR + name; //sub-group of some (non-root) group
   }
   else
     return GetPath(item);
@@ -893,7 +899,7 @@ void EditTreeLabel(wxTreeCtrl* tree, const wxTreeItemId& id)
   wxTextCtrl* edit = tree->EditLabel(id);
   if (edit) {
     wxTextValidator val(wxFILTER_EXCLUDE_CHAR_LIST);
-    const wxChar* dot = wxT(".");
+    const wxChar* dot = GROUP_SEL_STR;
     val.SetExcludes(wxArrayString(1, &dot));
     edit->SetValidator(val);
     edit->SelectAll();
@@ -904,7 +910,7 @@ void TreeCtrl::OnAddGroup(wxCommandEvent& WXUNUSED(evt))
 {
   wxCHECK_RET(IsShown(), wxT("Group can only be added while in tree view"));
   wxTreeItemId parentId = GetSelection();
-  wxString newItemPath = (!parentId || !parentId.IsOk() || parentId == GetRootItem() || !ItemIsGroup(parentId))? wxString(_("New Group")): GetItemGroup(parentId) + wxT(".") + _("New Group");
+  wxString newItemPath = (!parentId || !parentId.IsOk() || parentId == GetRootItem() || !ItemIsGroup(parentId))? wxString(_("New Group")): GetItemGroup(parentId) + GROUP_SEL_STR + _("New Group");
   if(Find(newItemPath, GetRootItem())) {
     wxMessageBox(_("\"") + _("New Group") + _("\" ") + _("name exists"), _("Double group name"), wxOK|wxICON_ERROR);
     return;
@@ -1007,6 +1013,214 @@ void TreeCtrl::OnKeyDown(wxTreeEvent& evt)
   evt.Skip();
 }
 
+TreeScrollTimer::TreeScrollTimer()
+{
+  m_owner = NULL;
+}
+
+void TreeScrollTimer::Notify()
+{
+  wxASSERT(m_owner);
+  m_owner->CheckScrollList();
+}
+
+TreeCollapseTimer::TreeCollapseTimer()
+{
+    m_owner = NULL;
+}
+
+void TreeCollapseTimer::Notify()
+{
+  wxASSERT(m_owner);
+  m_owner->CheckCollapseEntry();
+}
+
+void TreeCtrl::CheckScrollList()
+{
+  if(m_isDragging) {
+    int width, height;
+    wxPoint mousePt = ScreenToClient(wxGetMousePosition());  // Find where the mouse is in relation to the (exited) pane
+    
+    GetSize(&width, &height);  // Store window dimensions
+    
+    if((mousePt.x < 0) || (mousePt.x >= width)) { // Not in same area
+      resetScrolling();
+      return;
+    }
+    
+    int ThumbVert = GetScrollThumb (wxVERTICAL);
+    int PosVert = GetScrollPos (wxVERTICAL);
+    int RangeVert = GetScrollRange (wxVERTICAL);
+    wxEventType commandType = wxEVT_NULL;
+    
+    if(mousePt.y <= 0) { // We're above the pane
+      StopAutoScrolling();
+      if(PosVert > 0) {
+        commandType = wxEVT_SCROLLWIN_PAGEUP;
+      }
+    }
+    else if(mousePt.y > height) {  // We're below the pane
+      StopAutoScrolling();
+      if((PosVert + ThumbVert) < RangeVert) {
+        commandType = wxEVT_SCROLLWIN_PAGEDOWN;
+      }
+    }
+    else if(mousePt.y <= m_upper_scroll_limit) {
+      if(PosVert > 0) {
+        commandType = wxEVT_SCROLLWIN_LINEUP;
+      }
+    }
+    else if(mousePt.y >= m_lower_scroll_limit) {
+      if((PosVert + ThumbVert) < RangeVert) {
+        commandType = wxEVT_SCROLLWIN_LINEDOWN;
+      }
+    }
+    
+    if(commandType != wxEVT_NULL) {
+      wxScrollWinEvent scrollEvent( commandType, GetId(), 0);
+      GetEventHandler()->ProcessEvent(scrollEvent);
+      m_scroll_timer.Start();
+    }
+    else {
+      resetScrolling();
+    }
+  }
+}
+
+void TreeCtrl::CheckCollapseEntry()
+{
+  if(! m_last_mice_item_in_drag_and_drop.IsOk()) return;
+  
+  int width, height;
+  wxPoint mousePt = ScreenToClient(wxGetMousePosition());  // Find where the mouse is in relation to the (exited) pane
+  
+  GetSize(&width, &height);  // Store window dimensions
+  
+  if((mousePt.x < 0) || (mousePt.x >= width) || (mousePt.y < 0) || (mousePt.y >= height)) { // Not in same area
+    m_last_mice_item_in_drag_and_drop = nullptr;
+    return;
+  }
+  
+  int flags = wxTREE_HITTEST_ONITEM;
+  wxTreeItemId currentItem = DoTreeHitTest(mousePt, flags);
+  bool itemSame = (currentItem == m_last_mice_item_in_drag_and_drop);
+  
+  if(itemSame) {
+    if(currentItem.IsOk() && ItemHasChildren(currentItem) && (currentItem != GetRootItem())) {
+      if(IsExpanded(currentItem)) {
+        Collapse(currentItem);
+        wxTreeItemId selectedItem = GetSelection();
+        if(selectedItem.IsOk()) UnselectAll();
+      }
+      else { // Is Collapsed
+        Expand(currentItem);
+        //scroll the last child of this node into visibility
+        EnsureVisible(GetLastChild(currentItem));
+        //but if that scrolled the parent out of the view, bring it back
+        EnsureVisible(currentItem);
+      }
+    }
+  }
+  else {
+    m_last_mice_item_in_drag_and_drop = nullptr;
+  }
+}
+
+void TreeCtrl::OnMouseMove(wxMouseEvent& event)
+{
+  if(m_isDragging && m_drag_item.IsOk())
+  {
+    int width, height;
+    wxPoint pt = ScreenToClient(wxGetMousePosition());
+  
+    GetSize(&width,  &height);
+    
+    if((pt.x >= 0) && (pt.x < width)) { // When inside of the windows area
+      int flags = wxTREE_HITTEST_ONITEM;
+      wxTreeItemId currentItem = DoTreeHitTest(pt, flags);
+      bool itemChanged = (currentItem != m_last_mice_item_in_drag_and_drop);
+      int ThumbVert = GetScrollThumb (wxVERTICAL);
+      int PosVert = GetScrollPos (wxVERTICAL);
+      int RangeVert = GetScrollRange (wxVERTICAL);
+        
+      if((PosVert > 0) && (pt.y <= m_upper_scroll_limit)) {
+        if(! m_scroll_timer.IsRunning()) {
+          m_scroll_timer.Start();
+        }
+      } else if(((PosVert + ThumbVert) < RangeVert) && (pt.y >= m_lower_scroll_limit)){
+        if(! m_scroll_timer.IsRunning()) {
+          m_scroll_timer.Start();
+        }
+      }
+      else if(m_scroll_timer.IsRunning()) {
+        resetScrolling();
+      }
+        
+      if(currentItem.IsOk()) {
+        wxString label = GetItemText(m_drag_item);
+        wxTreeItemId si;
+        // Set Cursor depending from kind of entry
+        if(ItemIsGroupOrRoot(currentItem) && ! IsDescendant(currentItem, m_drag_item) && (::wxGetKeyState(WXK_SHIFT) || ! ExistsInTree(currentItem, tostringx(label), si))) {
+          SetCursor(wxNullCursor);
+        }
+        else {
+          SetCursor(wxCURSOR_NO_ENTRY);
+        }
+        // Check Collapse or Expand when taying for longer time over group entry
+        if(m_last_mice_item_in_drag_and_drop == nullptr) {
+          if(ItemHasChildren(currentItem)) {
+            m_last_mice_item_in_drag_and_drop = currentItem;
+            m_collapse_timer.Start();
+          }
+        } else if(itemChanged) {
+          if(m_collapse_timer.IsRunning())
+            m_collapse_timer.Stop();
+          if(ItemHasChildren(currentItem)) {
+            m_last_mice_item_in_drag_and_drop = currentItem;
+            m_collapse_timer.Start();
+          }
+        }
+        else { // Still the same, wait for timer expiry
+          // Do nothing
+        }
+      }
+      else {
+        if(m_last_mice_item_in_drag_and_drop)
+          SetCursor(wxNullCursor);
+        if(m_collapse_timer.IsRunning())
+          m_collapse_timer.Stop();
+        m_last_mice_item_in_drag_and_drop = nullptr;
+      }
+    }
+    else {
+      if(m_drag_image) {
+        m_drag_image->Hide();
+      }
+      if(m_last_mice_item_in_drag_and_drop || m_scroll_timer.IsRunning()) { // left or right from tree window, reset scrolling
+        resetScrolling();
+        resetDragItems();
+      }
+    }
+    
+    if((pt.x >= 0) && (pt.x < width) && (pt.y >= 0) && (pt.y < height)) {
+      StopAutoScrolling();
+      if(m_drag_image) {
+        m_drag_image->Hide();
+        m_drag_image->Move(pt);
+        m_drag_image->Show();
+      }
+      if(m_had_been_out) {
+        SetCursor(wxNullCursor);  // TO BE FIXED: This cursor is not visible because macOS means the cursor is not on the current windows area
+        m_had_been_out = false;
+      }
+    }
+    else
+      m_had_been_out = true;
+  }
+  
+  wxTreeCtrl::OnMouse(event);
+}
+
 void TreeCtrl::OnBeginDrag(wxTreeEvent& evt)
 {
   if (m_core.IsReadOnly() || !IsSortingGroup()) {
@@ -1018,14 +1232,49 @@ void TreeCtrl::OnBeginDrag(wxTreeEvent& evt)
 
   if (item.IsOk() && (GetRootItem() != item)) {
     m_drag_item = item;
+    m_had_been_out = false;
     SetWindowStyle(m_style & ~wxTR_HIDE_ROOT);
     evt.Allow();
     markDragItem(item);
     Refresh();
+    resetDragItems(true);
+    m_drag_image = new wxDragImage(*this, item);
+    if(m_drag_image) {
+      wxPoint mousePt = ScreenToClient(wxGetMousePosition());
+      wxRect rect;
+      GetBoundingRect(item, rect, true);
+      if(! m_drag_image->BeginDrag(mousePt - rect.GetLeftTop(), this, true)) {
+        pws_os::Trace(L"BeginDrag failed");
+        wxDELETE(m_drag_image);
+      }
+      else {
+        ReleaseMouse(); // Must be released because BeginDrag() will Capture Mouse, but the code following event BeginDragging() will do as well
+        m_drag_image->Show();
+      }
+    }
     return;
   }
 
   evt.Skip();
+}
+
+void TreeCtrl::resetScrolling()
+{
+  StopAutoScrolling();
+  if(m_scroll_timer.IsRunning()) m_scroll_timer.Stop();
+}
+
+void TreeCtrl::resetDragItems(bool initSize)
+{
+  if(m_scroll_timer.IsRunning()) m_scroll_timer.Stop();
+  if(m_collapse_timer.IsRunning()) m_collapse_timer.Stop();
+  m_last_mice_item_in_drag_and_drop = nullptr;
+  if(initSize) {
+    int width, height;
+    GetSize(&width,  &height);
+    m_lower_scroll_limit = height - GetCharHeight();
+    m_upper_scroll_limit = GetCharHeight();
+  }
 }
 
 void TreeCtrl::markDragItem(const wxTreeItemId itemSrc, bool markIt)
@@ -1047,14 +1296,41 @@ void TreeCtrl::OnEndDrag(wxTreeEvent& evt)
 {
   bool makeCopy = ::wxGetKeyState(WXK_CONTROL);
   bool doOverride = ::wxGetKeyState(WXK_SHIFT);
+  int width, height;
+  wxPoint mousePt = ScreenToClient(wxGetMousePosition());  // Find where the mouse is in relation to the (exited) pane
+  GetSize(&width, &height);  // Store window dimensions
+  int flags = wxTREE_HITTEST_ONITEM;
+  wxTreeItemId currentItem = DoTreeHitTest(mousePt, flags); // Check current before removing root from screen
+  resetDragItems();
   // Restore Hiden Root
   SetWindowStyle(m_style);
   markDragItem(m_drag_item, false);
+  if(m_drag_image) {
+    m_drag_image->Hide();
+    CaptureMouse(); // EndDrag() will release the Mouse, but before call the mouse had been released from event handler
+    m_drag_image->EndDrag();
+    wxDELETE(m_drag_image);
+  }
   Refresh();
-  // Do not drag on itself
-  if ((m_drag_item != NULL) && (m_drag_item == GetSelection())) {
+  
+  wxTreeItemId selectedItem = GetSelection();
+  if(selectedItem.IsOk()) UnselectAll();
+  
+  if((mousePt.x < 0) || (mousePt.x >= width) || (mousePt.y < 0) || (mousePt.y >= height)) { // Not in same area
+    evt.Skip();
+    m_drag_item = nullptr;
+    return;
+  }
+  
+  if (m_drag_item != nullptr) { // Drag and drop started
 
     wxTreeItemId itemDst = evt.GetItem();
+    
+    if(itemDst.IsOk() && (currentItem != itemDst)) { // Mice is not over destination item, skip dragging
+      evt.Skip();
+      m_drag_item = nullptr;
+      return;
+    }
 
     if(itemDst && itemDst.IsOk() &&
        (itemDst != m_drag_item) &&
@@ -1062,15 +1338,15 @@ void TreeCtrl::OnEndDrag(wxTreeEvent& evt)
        ((GetRootItem() != itemDst) || (GetRootItem() != GetItemParent(m_drag_item)))) { // Do not Drag and Drop on its own
       if(GetRootItem() != itemDst) {
         if(! ItemIsGroup(itemDst)) {                   // Only group may be destination
-          wxMessageBox(_("Drag and Drop failed"), _("Destination is no group"), wxOK|wxICON_ERROR);
+          wxMessageBox(_("Destination is not a group"), _("Drag and Drop failed"), wxOK|wxICON_ERROR);
           evt.Skip();
-          m_drag_item = NULL;
+          m_drag_item = nullptr;
           return;
         }
         else if(IsDescendant(itemDst, m_drag_item)) {  // Do not drag and drop into the moved tree
-          wxMessageBox(_("Drag and Drop failed"), _("Destination cannot be inside source tree"), wxOK|wxICON_ERROR);
+          wxMessageBox(_("Destination cannot be inside source tree"), _("Drag and Drop failed"), wxOK|wxICON_ERROR);
           evt.Skip();
-          m_drag_item = NULL;
+          m_drag_item = nullptr;
           return;
         }
       }
@@ -1112,7 +1388,7 @@ void TreeCtrl::OnEndDrag(wxTreeEvent& evt)
         StringX sxNewPath;
         
         if(GetRootItem() != itemDst) {
-          sxNewPath = tostringx(GetItemGroup(itemDst) + wxT(".") + label);
+          sxNewPath = tostringx(GetItemGroup(itemDst) + GROUP_SEL_STR + label);
         }
         else {
           sxNewPath = tostringx(label);
@@ -1121,8 +1397,8 @@ void TreeCtrl::OnEndDrag(wxTreeEvent& evt)
           wxTreeItemId si;
           if(ExistsInTree(itemDst, tostringx(label), si)) {
             evt.Veto();
-            wxMessageBox(_("Same group name exists (use Shift to overrule)"), _("Double group name"), wxOK|wxICON_ERROR);
-            m_drag_item = NULL;
+            wxMessageBox(_("Duplicate group name (use Shift to overrule)"), _("Double group name"), wxOK|wxICON_ERROR);
+            m_drag_item = nullptr;
             return;
           }
         }
@@ -1141,10 +1417,14 @@ void TreeCtrl::OnEndDrag(wxTreeEvent& evt)
           wxTreeCtrl::SelectItem(newItem);
       }
     }
+    else {
+      if(m_drag_item.IsOk())
+        wxTreeCtrl::SelectItem(m_drag_item);
+    }
   }
 
   evt.Skip();
-  m_drag_item = NULL;
+  m_drag_item = nullptr;
 }
 
 bool TreeCtrl::IsDescendant(const wxTreeItemId itemDst, const wxTreeItemId itemSrc)
@@ -1224,7 +1504,7 @@ void TreeCtrl::FinishAddingGroup(wxTreeEvent& evt, wxTreeItemId groupItem)
     wxString groupName = evt.GetLabel();
     for (wxTreeItemId parent = GetItemParent(groupItem);
                       parent != GetRootItem(); parent = GetItemParent(parent)) {
-      groupName = GetItemText(parent) + wxT(".") + groupName;
+      groupName = GetItemText(parent) + GROUP_SEL_STR + groupName;
     }
     StringX sxGroup = tostringx(groupName);
 
@@ -1358,7 +1638,7 @@ void TreeCtrl::ExtendCommandCopyGroup(MultiCommands* pmCmd, wxTreeItemId itemSrc
     const StringX label = tostringx(GetItemText(ti));
     
     if(ItemIsGroup(ti)) {
-      ExtendCommandCopyGroup(pmCmd, ti, sxNewPath + wxT(".") + label, checkName);
+      ExtendCommandCopyGroup(pmCmd, ti, sxNewPath + GROUP_SEL_STR + label, checkName);
     }
     else {
       CItemData *dataSrc = TreeCtrl::GetItem(ti);