+++ docs/ReleaseNotes.md
@@ -12,6 +12,7 @@ In the following, SFxxxx refers to Bug Reports, Feature Requests and Service Req
 
 New features in 3.55
 --------------------
+* [SF892](https://sourceforge.net/p/passwordsafe/feature-requests/892/) Added tooltip to copy password button in Add/Edit and Password Policy windows.
 * [SF525](https://sourceforge.net/p/passwordsafe/support-requests/525/) Removed limitation on encrytping/decrypting files larger than 2GB.
 
 Bugs fixed in 3.54.01